added full RFID to Unity code, added some ending descriptions

+++ Scripts_0.2/HumanActivity.cs
@@ -38,7 +38,7 @@ public class HumanActivity
     public string ActivityClass {get=>activityClass;set=>activityClass = value;}
     public string ActivityGroup {get=>activityGroup;set=>activityGroup=value;}
     public int Score {get=>score; set=>score=value;}
-    public int CarbonEmission{get=>carbonEmission;set=>carbonEmission=value;}
+    public float CarbonEmission{get=>carbonEmission;set=>carbonEmission=value;}
 
     public HumanActivity() {
